patchtools: change missing commits to missing fixes
This is more aligned with what they are called upstream.
Also added the upstream definition of Fixes to the help text and
the informational messaging for selecting the remote repo/branch
in git_remote_seek.

In git_remote_seek, made the call to select_rembranch depend on the
return from select_remrepo. If the user pressed 'q' in that routine,
there is no need to call select_rembranch.