patchtools: patchreview: refactoring and new funcionality
Deleted a lot of old and useless comments.

Added b_verbose, b_mrcomments, and b_seekfixes to set/get_boolean
functions.

Added functionality to automatically check_fixes and put the option
in the config menu to toggle.

Implemented a check of the user's config file with the config file
in the lib directory. If there are new fields, then the user's
config file will have the new fields added with their default
values. All the other user config variables remain the same.

Added a lot of functionality to the command line operation as
well as a command line help.

modified:   patchreview