patchtools: WIP add option for external upstream repo
The new KWF enforces a separate directory for each repe, e.g. RHEL7,
RHEL8, RHEL9, etc. In the old workflow, these were remotes that were
added to the upstream repo.

In the new KWF, the upstream repo was added as a remote to the RHEL
repo. However, this meant replicating the upstream as a remote for
each RHEL distro.

It seems better to put the upstream repo in a separate directory so
it does not have to be replicated as a remote for each distro.

The option is still there to create the upstream repo as a remote
in each distro directory.

modified:   lib/patch-common.source
modified:   lib/patchreview.conf
modified:   lib/ui.source
modified:   patchreview