patchtools: ui_waitonproc_tty for search of RHEL repo for missing fixes
If the missing fix wasn't intentionally omitte, and if it is not
included in the current series, and it is nowhere to be found in the
RHEL repo/branch, then it is really missing. The search of the RHEL
repo can take some time, so print dots while we do that.