mkworkdir: create a /work mount point and corresponding logical vol
This assumes xfs. Do not run this script with any other fs.

Removes and recreates the home logical volume with a smaller size
input by the user as prompted.

Creates a work logical volume with a /work mount point with a size
equal to the space freed by recreating home with less space.