patchtools: patch-mrutilities: patchcount again
Seems that all kinds of things can obscure the patchcount when there
are dependencies, stale HEADs, etc.

So this change counts only the commits in the actual MR, rather than
trying to 'git log --reverse' or anything like that.

It's actually much simpler, even though it does depend on the lab
utility.