patchtools: gitutilities: git_comment2commitsfile
Scan all the tokens in the RHEL subject line. If the developer
created patches with summaries that are too different than the
corresponding upstream summaries, then it will break.

Need to revisit this to have the user view the differences, rather
than just skipping it. It's a tough call. I don't want to annoy
the user by presenting a lot of choices, but I don't want to drop
commits because the developer munged the summary lines beyond the
script's abilities to discern it.