lib/gitutilities: git_comment2commitsfile improvements
Vast improvement in the logic. Pipe the out put of grep for commit
hashes to sort -u in order to squeeze out duplicates. If dups are
not squeezed out, an unnecessary call is made to validate the patch
subject line.

Add logic for the grep to omit any reverted commits by using the -v
switch and the -i switch for case insensitive.