patchtools: patcmp: fix backup and patch substitution logic
Because there is a hard maximum index set to the number of patches, or
the number of mismatches in mismatch mode, there was an indexing problem
when reaching the maxindex or last_mismatch value. This was especially
painful when replacing a commit hash, because you'd always get the patch
BEFORE the one you want, if it were the last one in the series or the
last one in the mismatch_array.

Special case those in the backup logic to fix this hideous bug.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>