patchtools: patchreview: Add M and m to ack/nack menu
User may want to review more MRs after ack/nack, so put that option
into the ack/nack menu, instead of making the user go all the way
back to the main menu to do it.

Fixed a minor logic bug in the main loop that came from implementing
the ShellCheck suggestions.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>