lib/gitutilities: make mrcomments log if needed
Also only look for the first instance of a commit hash when seeking
fixes in the usdir/us-commits.log. While this eliminates false
positives, it does not catch using the same upstream commit hash in
two different backported upstream commits. I'll have to rethink
this a bit, but this fix will do for now.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>