patchtools: cherp: introduce stop hash and cherp apply
The stophash variable tells cherp to stop cherrypicking before
the upstream commit identified by stophash. This is useful for
stopping processing before encountering something like a tree-wide
commit, or any other commit that requires modification before
applying.

Once the commit has been massaged and formatted so that it will
apply cleanly, cherp apply is used to create the commit log and
advance the processing to the next upstream commit.

The backport.log is also updated so that each commit that is
successfully applied will have a leading + prepended to the
corresponding line in the backport.log file.

The man text and config file have been updated accordingly.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>