lib/ui: add a color and change the WRN color.
Also replaced single quotes with double quotes in the ANSI escape
sequences. If they contain any operations, they won't be performed
with the single quotes.

Added LIL (lilac) and BLIL (bold lilac), a bright magenta. The BLIL
will replace the BRED for the WRN color on dark backgrounds. BRED
was just too hard to see on a dark background.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>