patchtools: mkbackportlog: downstream commit must be supplied
Forgot that you must have a downstream commit to be able to search
downstream for upstream commits that have already been backported.

Did a lot of updating on the help text. Still need to write a manual
for the menu-driven interface. I'd like to think it's self-documenting,
but this is a daunting paradigm for the new user

Signed-off-by: Tony Camuso <tcamuso@redhat.com>