keypress: decided to do this in C instead
Python is just too screwy. Trying to install libraries is just so
painful when things go awry.

The libraries for C are usually very handy, and it only takes a
few seconds at most  to compile a small source file into a native
executable to be used with bash scripts.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>