lib/ui: add keypress and keypress test functionality
Adding some options in this script that require Alt-Shift keystrokes
to execute. These options won't be available without the natively compiled
keypress utility in this repo. These options can do things like delete
files or reset git heads, etc. The alt-shift keystrokes provide a level
of safety from inadvertent keystrokes modifying your development state.

The get_keypress function is added to acquire Alt-Shift keystrokes. If
the keypress executable is not present, and cannot or will not be
compiled, then the call to get_keypress will failover to getchar.

However, the intent is not to even call get_keypress unless the keypress
executable is present, because we don't want to enable dangerous options
without the safety of an Alt-Shift keystroke.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>