patchtools: cherp: better tracking of origin head, etc
Get the name of the origin's branch by using 'git remote show origin'
This takes enough time to warrant a ui_waitonproc call, which also
requires a tempfile, using mktemp.

After using cut, the branch name still has a leading space, so that
has to be trimmed.

The original downstream head is then determined by using..
'git merge-base HEAD <origin_branch_name>. This is a much better way
to determine where the current branch diverged from the origin main
or master branch.

In all, this is a much more deterministic way to get the current
downstream state of the main and current branches.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>