patchtools: wcherp: WIP - tmux wrapper for cherp
Splits the window vertically or horizontally, selectable by a dash
option. The default is vertical. cherp will run in one pane, and
vim will run in the other pane. The backports commit file is what
will be shown in the vim window, and vim will be invoked so that
it follows any updates.

In a vertical split, cherp will be on the left and vim on the right.
In a horizontal split, cherp will be on top, and vim on the bottom.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>