patchtools: cherp: use patch instead of git apply
Wasn't able to find a way for 'git apply' to apply commits when the
patch contained patches for a file that did not exist in the index.
The 'patch' application did this.

The reason for having alternatives to git cherry-pick, is that it
can be slower and have housekeeping problems, e.g. merge.renameLimit=1.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>