patchtools: look through whole commit log for upstream commits
The -m3 switch to the grep that looks for upstream commits was
limiting the search to the first three lines of the commit log.
Consequently, if there were any intervening lines, e.g. a conflcts
stanza, the upstream commit would be missed.

Possible fallout is that there would be commits with 40 char hashes
in the conflicts stanza, or elsewhere in the commit log, and the
user will be presented with a list of them to choose the one s/he
wants.

This is somewhat of a corner case, so pain would be limited.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>