mkworkdir-noq: WIP: make this script more intelligent
Need to include mapper and new UUID fields to update the fstab.
Use a function to update the fstab, so we don't have to do the
same thing more than once for each resized logical volume.
Call systemctl daemon-reload to refresh the fstab in memory.

Need to test it, first, and then do the same for mkworkdir.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>