init-my-stuff: add newuser logic"
The newuser was hardcoded to be tcamuso. Changed that to a query
for newuser/password, but you can't query in a non-interactive
script, which this is.

The query was removed and the newuser was added as an argument.
The newuser password will have to be created manually from the
root prompt after the install.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>