patchtools: getup: fix no orighead bug
If the dnorighead global was empty, the double quotes around it
made 'git log' look for a null commit. Remove the double quotes,
and it simply becomes a 'git log' with no commit when dnorighead
is null.

Also fixed error signal caused by trying to access a nonexistent
file. New session must be run, but if it has not, then there is
no file of backport commits.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>