patchtools: gorp: final touches for v12.7
getup was renamed gorp, so there's a big deletion of that file.

gorp got a new function, git_verify, which consists of a call to
git rev-parse to validate whether a given hash is a valid commit.
In flag_conflict, we had to test whether the conflicted patch file
was already created for the given hash. This was to prevent over-
writing the file, just in case it had been edited. That meant we
had to add a new field to lib/gorp.conf.

Changed vanilla git apply to git apply --reject. The .rej files are
very helpful for sorting conflicts.

There was also a recursive call to show_status > flag_conflict >
show_status. Easy fix.

The biggest changes were made in man/gorp.pgman. I think I have it
pretty much upto date, though I wouldn't be surprised if there were
a typo or cut-and-paste error here and there.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>