patchtools: extup: speed up with git native and awk
Also removed the pushd/popd pairs to go in and out of the upstream
repo directory. This is no longer needed, since we have that repo
mirrored in a parallel directory. This allows us to maintain just
one full-weight upstream repo and one lightweight mirror for all
the downstream repos to reference.

TODO: Update the help with details about setting up the mirrored
      repo.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>