patchtools: gort: WIP: Add option enter_newhead
TODO: Need to debug stage_commit to determine why the Jira and Upstream
      status stanzas disappear after handling a conflict.
      set -x and set +x are in place

enter_newhead allows the user to reset the head. This is important when
switching branches. Maybe it should be called from select_gitbranch as
well as being an Alt-Shift menu option.

Added quit option in select_gitbranch, to preclude having to issue git
reset, etc.

Added init_newhead procedure to obviate repeating the same newhead
sequence of commands in a number of places.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>